fix(worker): move worker dir from core/ to app/
