+++ src/app/worker/hooks.py
@@ -4,7 +4,7 @@ from typing import Any
 
 import uvloop
 
-from ..db.database import async_engine
+from ..core.db.database import async_engine
 
 asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
