+++ src/app/worker/ranges.py
@@ -9,15 +9,15 @@ from src.app.crud.crud_ranges import create_deployed_range, delete_deployed_rang
 from src.app.enums.range_states import RangeState
 from src.app.schemas.user_schema import UserID
 
-from ...core.cdktf.ranges.range_factory import RangeFactory
-from ...crud.crud_users import get_decrypted_secrets, get_user_by_id
-from ...schemas.range_schemas import (
+from ..core.cdktf.ranges.range_factory import RangeFactory
+from ..core.db.database import get_db_session_context
+from ..crud.crud_users import get_decrypted_secrets, get_user_by_id
+from ..schemas.range_schemas import (
     BlueprintRangeSchema,
     DeployedRangeSchema,
     DeployRangeSchema,
 )
-from ...utils.job_utils import track_job_status
-from ..db.database import get_db_session_context
+from ..utils.job_utils import track_job_status
 
 asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
